+++ api/questions.js
@@ -96,6 +96,26 @@ router.post(
   [requireCourseStaffForQueueForQuestion, requireQuestion, failIfErrors],
   async (req, res, _next) => {
     const { question } = res.locals
+
+    if (question.beingAnswered) {
+      // Forbid someone else from taking over this question
+      res.status(403).send('Another user is already answering this question')
+      return
+    }
+
+    // Verify that this user isn't currently answering another question
+    const otherQuestions = await Question.find({
+      where: {
+        answeredById: res.locals.userAuthn.id,
+        dequeueTime: null,
+      },
+    })
+
+    if (otherQuestions !== null) {
+      res.status(403).send('You are already answering this question')
+      return
+    }
+
     modifyBeingAnswered(question, true)
     question.answeredById = res.locals.userAuthn.id
     await question.save()
@@ -131,10 +151,24 @@ router.post(
   async (req, res, _next) => {
     const data = matchedData(req)
 
+    // Temporary, easy fix to avoid having to rename enums
+    // TODO Fix this garbage
+    let mappedPreparedness = data.preparedness
+    switch (data.preparedness) {
+      case 'bad':
+        mappedPreparedness = 'not'
+        break
+      case 'good':
+        mappedPreparedness = 'well'
+        break
+      default:
+        break
+    }
+
     const { question } = res.locals
     question.answerFinishTime = new Date()
     question.dequeueTime = new Date()
-    question.preparedness = data.preparedness
+    question.preparedness = mappedPreparedness
     question.comments = data.comments
     question.answeredById = res.locals.userAuthn.id